--- frontend/src/accounts/Accounts.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { LoginForm } from "./AccountsForms";
-
-export const Login = () => (
-    <LoginForm/>
-)
