+++ frontend/src/accounts/LoginForm.js
@@ -7,11 +7,11 @@ import Col from 'react-bootstrap/Col';
 import { Form as BootstrapForm } from 'react-bootstrap';
 import * as Yup from "yup";
 import { useCookies } from 'react-cookie';
-import { TextInput } from '.././components/Form.js';
+import { TextInput } from '../components/Form.js';
 import { AuthContext } from "./AccountsReducer";
 import { loadUser, setAuthToken } from "./AccountsUtils";
 
-export const LoginForm = () => {
+const Login = () => {
   const { state, dispatch } = useContext(AuthContext);
   const [cookies, setCookie, removeCookie] = useCookies(['token']);
 
@@ -89,3 +89,5 @@ export const LoginForm = () => {
     </Fragment>
   )
 }
+
+export default Login;
