+++ frontend/src/shelter/BuildingForm.js
@@ -1,0 +1,136 @@
+import React, { useContext, useEffect, useState } from 'react';
+import axios from "axios";
+import { navigate, useQueryParams } from 'raviger';
+import { Formik } from 'formik';
+import { Form as BootstrapForm, Button, ButtonGroup, Card } from "react-bootstrap";
+import { TextInput } from '../components/Form';
+import * as Yup from 'yup';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faArrowAltCircleLeft } from '@fortawesome/free-solid-svg-icons';
+import { AuthContext } from "../accounts/AccountsReducer";
+
+// Regex validators.
+const nameRegex = /^[a-z0-9 ,.'-]+$/i
+
+export const BuildingForm = ({id}) => {
+
+  const { state, dispatch } = useContext(AuthContext);
+
+  // Identify any query param data.
+  const [queryParams] = useQueryParams();
+  const {
+    shelter_id = ''
+  } = queryParams;
+
+  // Initial Building data.
+  const [data, setData] = useState({
+    name: '',
+    description: '',
+    shelter: shelter_id,
+  });
+
+  // Hook for initializing data.
+  useEffect(() => {
+    let source = axios.CancelToken.source();
+    if (id) {
+      const fetchBuildingData = async () => {
+        // Fetch ServiceRequest data.
+        await axios.get('/shelter/api/building/' + id + '/', {
+          cancelToken: source.token,
+        })
+        .then(response => {
+          setData(response.data);
+        })
+        .catch(error => {
+          console.log(error.response);
+        });
+      };
+      fetchBuildingData();
+    }
+
+    // Cleanup.
+    return () => {
+      source.cancel();
+    };
+  }, [id]);
+
+  return (
+    <>
+      <Formik
+        initialValues={data}
+        enableReinitialize={true}
+        validationSchema={Yup.object({
+          name: Yup.string()
+            .max(50, 'Must be 50 characters or less')
+            .matches(nameRegex, "Name is not valid")
+            .required('Required'),
+          description: Yup.string()
+            .max(400, 'Must be 400 characters or less'),
+        })}
+        onSubmit={(values, { setSubmitting }) => {
+          setTimeout(() => {
+            if (id) {
+              axios.put('/shelter/api/building/' + id + '/', values)
+              .then(function() {
+                if (state.prevLocation) {
+                  navigate(state.prevLocation);
+                }
+                else {
+                  navigate('/shelter/building/' + id);
+                }
+              })
+              .catch(e => {
+                console.log(e);
+              });
+            }
+            else {
+              axios.post('/shelter/api/building/', values)
+              .then(function() {
+                if (state.prevLocation) {
+                  navigate(state.prevLocation);
+                }
+                else {
+                  navigate('/shelter/' + shelter_id);
+                }
+              })
+              .catch(e => {
+                console.log(e);
+              });
+            }
+            setSubmitting(false);
+          }, 500);
+        }}
+      >
+        {props => (
+          <Card border="secondary" className="mt-5">
+          <Card.Header as="h5" className="pl-3"><span style={{cursor:'pointer'}} onClick={() => window.history.back()} className="mr-3"><FontAwesomeIcon icon={faArrowAltCircleLeft} size="lg" inverse /></span>{!id ? "New" : "Update"} Building</Card.Header>
+          <Card.Body>
+            <BootstrapForm noValidate>
+              <BootstrapForm.Row>
+                <TextInput
+                  xs="5"
+                  type="text"
+                  label="Name*"
+                  name="name"
+                  id="name"
+                />
+                <TextInput
+                  xs="5"
+                  type="text"
+                  label="Description"
+                  name="description"
+                />
+              </BootstrapForm.Row>
+            </BootstrapForm>
+          </Card.Body>
+          <ButtonGroup size="lg">
+            <Button type="submit" onClick={() => { props.submitForm()}}>Save</Button>
+          </ButtonGroup>
+        </Card>
+        )}
+      </Formik>
+    </>
+  );
+};
+
+export default BuildingForm;
