+++ frontend/src/shelter/ShelterDetails.js
@@ -11,7 +11,7 @@ import History from '../components/History';
 import Header from '../components/Header';
 import noImageFound from '../static/images/image-not-found.png';
 
-export function ShelterDetails({id}) {
+function ShelterDetails({id}) {
 
   const [data, setData] = useState({
     name: '',
@@ -208,245 +208,4 @@ export function ShelterDetails({id}) {
   );
 };
 
-export function BuildingDetailsTable({id}) {
-
-  const [data, setData] = useState({name:'', description: '', shelter: null, shelter_name:'', rooms:[], action_history:[]});
-
-  // Hook for initializing data.
-  useEffect(() => {
-    let source = axios.CancelToken.source();
-    const fetchBuildingData = async () => {
-      // Fetch Building Details data.
-      await axios.get('/shelter/api/building/' + id + '/', {
-        cancelToken: source.token,
-      })
-      .then(response => {
-        setData(response.data);
-      })
-      .catch(e => {
-        console.log(e);
-      });
-    };
-    fetchBuildingData();
-  }, [id]);
-
-  return (
-    <>
-    <Header>
-      Building Details
-      <OverlayTrigger
-        key={"edit-building"}
-        placement="bottom"
-        overlay={
-          <Tooltip id={`tooltip-edit-building`}>
-            Update building
-          </Tooltip>
-        }
-      >
-        <Link href={"/shelter/building/edit/" + id}><FontAwesomeIcon icon={faEdit} className="ml-1" inverse /></Link>
-      </OverlayTrigger>
-    </Header>
-    <hr/>
-    <Card className="border rounded d-flex" style={{width:"100%"}}>
-      <Card.Body>
-        <Card.Title>
-          <h4>Information
-            <OverlayTrigger key={"assign"} placement="top" overlay={<Tooltip id={`tooltip-assign`}>Assign animals to rooms</Tooltip>}>
-              <Link href={"/shelter/" + data.shelter + "/assign"}><FontAwesomeIcon className="ml-1" icon={faWarehouse} inverse/></Link>
-            </OverlayTrigger>
-          </h4>
-        </Card.Title>
-        <hr/>
-        <ListGroup variant="flush" style={{marginTop:"-13px", marginBottom:"-13px"}}>
-          <ListGroup.Item>
-            <b>Name:</b> {data.name}
-          </ListGroup.Item>
-          {data.description ? <ListGroup.Item>
-            <b>Description: </b>{data.description}
-          </ListGroup.Item> : ""}
-          <ListGroup.Item>
-            <b>Shelter:</b> {data.shelter_name}
-            <OverlayTrigger
-              key={"shelter-details"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-shelter-details`}>
-                  Shelter details
-                </Tooltip>
-              }
-            >
-              <Link href={"/shelter/" + data.shelter}><FontAwesomeIcon icon={faClipboardList} className="ml-1" inverse /></Link>
-            </OverlayTrigger>
-          </ListGroup.Item>
-        </ListGroup>
-      </Card.Body>
-    </Card>
-    <Card className="mt-3 border rounded d-flex">
-      <Card.Body>
-        <Card.Title>
-          <h4>Rooms
-            <OverlayTrigger
-              key={"add-room"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-add-room`}>
-                  Add a room
-                </Tooltip>
-              }
-            >
-            <Link href={"/shelter/building/room/new?building_id=" + id}><FontAwesomeIcon icon={faPlusSquare} className="ml-1" inverse /></Link>
-          </OverlayTrigger>
-          </h4>
-        </Card.Title>
-        <hr/>
-        <span className="d-flex flex-wrap align-items-end">
-          {data.rooms.map(room => (
-            <Card key={room.id} className="border rounded mr-3" style={{width:"100px", height:"100px"}}>
-              <Card.Text className="text-center mb-0">
-                {room.name}
-                <OverlayTrigger
-                  key={"room-details"}
-                  placement="top"
-                  overlay={
-                    <Tooltip id={`tooltip-room-details`}>
-                      Room details
-                    </Tooltip>
-                  }
-                >
-                  <Link href={"/shelter/room/" + room.id}><FontAwesomeIcon icon={faClipboardList} className="ml-1" inverse /></Link>
-                </OverlayTrigger>
-              </Card.Text>
-              <Card.Text className="text-center mb-0">
-                {room.animal_count} Animals
-              </Card.Text>
-            </Card>
-          ))}
-        </span>
-      </Card.Body>
-    </Card>
-    <History action_history={data.action_history} />
-    </>
-  );
-};
-
-export function RoomDetailsTable({id}) {
-
-  const [data, setData] = useState({name:'', description:'', building_name: '', shelter_name:'', shelter: null, animals:[], action_history:[]});
-
-  // Hook for initializing data.
-  useEffect(() => {
-    let source = axios.CancelToken.source();
-    const fetchRoomData = async () => {
-      // Fetch Room Details data.
-      await axios.get('/shelter/api/room/' + id + '/', {
-          cancelToken: source.token,
-      })
-      .then(response => {
-          setData(response.data);
-      })
-      .catch(e => {
-          console.log(e);
-      });
-    };
-    fetchRoomData();
-  }, [id]);
-
-  return (
-    <>
-    <Header>
-      Room Details
-      <OverlayTrigger
-        key={"edit-room"}
-        placement="bottom"
-        overlay={
-          <Tooltip id={`tooltip-edit-room`}>
-            Update room
-          </Tooltip>
-        }
-      >
-        <Link href={"/shelter/room/edit/" + id}><FontAwesomeIcon icon={faEdit} className="ml-1" inverse /></Link>
-      </OverlayTrigger>
-    </Header>
-    <hr/>
-    <Card className="border rounded d-flex" style={{width:"100%"}}>
-      <Card.Body>
-        <Card.Title>
-          <h4>Information
-            <OverlayTrigger key={"assign"} placement="top" overlay={<Tooltip id={`tooltip-assign`}>Assign animals to rooms</Tooltip>}>
-              <Link href={"/shelter/" + data.shelter + "/assign"}><FontAwesomeIcon className="ml-1" icon={faWarehouse} inverse/></Link>
-            </OverlayTrigger>
-          </h4>
-        </Card.Title>
-        <hr/>
-        <ListGroup variant="flush" style={{marginTop:"-13px", marginBottom:"-13px"}}>
-          <ListGroup.Item>
-            <b>Name:</b> {data.name}
-          </ListGroup.Item>
-          {data.description ? <ListGroup.Item>
-            <b>Description: </b>{data.description}
-          </ListGroup.Item> : ""}
-          <ListGroup.Item>
-            <b>Building:</b> {data.building_name}
-            <OverlayTrigger
-              key={"building-details"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-building-details`}>
-                  Building details
-                </Tooltip>
-              }
-            >
-              <Link href={"/shelter/building/" + data.building}><FontAwesomeIcon icon={faClipboardList} className="ml-1" inverse /></Link>
-            </OverlayTrigger>
-          </ListGroup.Item>
-          <ListGroup.Item>
-            <b>Shelter:</b> {data.shelter_name}
-            <OverlayTrigger
-              key={"shelter-details"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-shelter-details`}>
-                  Shelter details
-                </Tooltip>
-              }
-            >
-              <Link href={"/shelter/" + data.shelter}><FontAwesomeIcon icon={faClipboardList} className="ml-1" inverse /></Link>
-            </OverlayTrigger>
-          </ListGroup.Item>
-        </ListGroup>
-      </Card.Body>
-    </Card>
-    <div className="row mb-2 mt-3" hidden={data.animals.length === 0}>
-      <div className="col-12 d-flex">
-        <Card className="mb-2 border rounded" style={{width:"100%"}}>
-          <Card.Body>
-            <Card.Title>
-              <h4 className="mb-0">Animals</h4>
-            </Card.Title>
-            <hr/>
-            <span className="d-flex flex-wrap align-items-end">
-            {data.animals.map(animal => (
-              <Card key={animal.id} className="mr-3" style={{border:"none"}}>
-                <ReactImageFallback style={{width:"151px", height:"151px", objectFit: "cover", overflow: "hidden"}} src={animal.front_image} fallbackImage={[animal.side_image, noImageFound]} />
-                <Card.Text className="text-center mb-0">
-                  {animal.name||"Unknown"}
-                  <Link href={"/animals/" + animal.id}> <FontAwesomeIcon icon={faClipboardList} inverse /></Link>
-                  <Link href={"/animals/edit/" + animal.id}> <FontAwesomeIcon icon={faEdit} inverse /></Link>
-                </Card.Text>
-                <Card.Text className="text-center mb-0">
-                  {animal.status}
-                </Card.Text>
-                <Card.Text className="text-center" style={{textTransform:"capitalize"}}>
-                  {animal.size} {animal.species}
-                </Card.Text>
-              </Card>
-            ))}
-            </span>
-          </Card.Body>
-        </Card>
-      </div>
-    </div>
-    <History action_history={data.action_history} />
-    </>
-  );
-};
+export default ShelterDetails;
